requirements.txt
customtkinter
darkdetect
loguru
Pillow
